+++ feature-libs/product-configurator/rulebased/occ/cpq/cpq-configurator-occ.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { ConfigModule } from '@spartacus/core';
 import { CPQ_CONFIGURATOR_ADD_TO_CART_SERIALIZER } from './converters/cpq-configurator-occ.converters';
 import { OccConfiguratorCpqAddToCartSerializer } from './converters/occ-configurator-cpq-add-to-cart-serializer';
-import { defaultOccCpqConfiguratorConfigFactory } from './default-occ-cpq-configurator-config';
+import { defaultOccCpqConfiguratorConfigFactory } from './default-occ-configurator-cpq-config';
 
 @NgModule({
   imports: [
