+++ feature-libs/product-configurator/rulebased/occ/cpq/default-occ-configurator-cpq-config.ts
